+++ tests/unit/sagemaker/image_uris/test_huggingface_llm.py
@@ -19,28 +19,39 @@ from tests.unit.sagemaker.image_uris import expected_uris, conftest
 
 LMI_VERSIONS = ["0.24.0"]
 HF_VERSIONS_MAPPING = {
-    "0.6.0": "2.0.0-tgi0.6.0-gpu-py39-cu118-ubuntu20.04",
-    "0.8.2": "2.0.0-tgi0.8.2-gpu-py39-cu118-ubuntu20.04",
-    "0.9.3": "2.0.1-tgi0.9.3-gpu-py39-cu118-ubuntu20.04",
-    "1.0.3": "2.0.1-tgi1.0.3-gpu-py39-cu118-ubuntu20.04",
-    "1.1.0": "2.0.1-tgi1.1.0-gpu-py39-cu118-ubuntu20.04",
-    "1.2.0": "2.1.1-tgi1.2.0-gpu-py310-cu121-ubuntu20.04",
-    "1.3.1": "2.1.1-tgi1.3.1-gpu-py310-cu121-ubuntu20.04",
+    "gpu": {
+        "0.6.0": "2.0.0-tgi0.6.0-gpu-py39-cu118-ubuntu20.04",
+        "0.8.2": "2.0.0-tgi0.8.2-gpu-py39-cu118-ubuntu20.04",
+        "0.9.3": "2.0.1-tgi0.9.3-gpu-py39-cu118-ubuntu20.04",
+        "1.0.3": "2.0.1-tgi1.0.3-gpu-py39-cu118-ubuntu20.04",
+        "1.1.0": "2.0.1-tgi1.1.0-gpu-py39-cu118-ubuntu20.04",
+        "1.2.0": "2.1.1-tgi1.2.0-gpu-py310-cu121-ubuntu20.04",
+        "1.3.1": "2.1.1-tgi1.3.1-gpu-py310-cu121-ubuntu20.04",
+        "1.3.3": "2.1.1-tgi1.3.3-gpu-py310-cu121-ubuntu20.04",
+    },
+    "inf2": {
+        "0.0.16": "1.13.1-optimum0.0.16-neuronx-py310-ubuntu22.04",
+        "0.0.17": "1.13.1-optimum0.0.17-neuronx-py310-ubuntu22.04",
+    },
 }
 
 
-@pytest.mark.parametrize("load_config", ["huggingface-llm.json"], indirect=True)
+@pytest.mark.parametrize(
+    "load_config", ["huggingface-llm.json", "huggingface-llm-neuronx.json"], indirect=True
+)
 def test_huggingface_uris(load_config):
     VERSIONS = load_config["inference"]["versions"]
+    device = load_config["inference"]["processors"][0]
+    backend = "huggingface-neuronx" if device == "inf2" else "huggingface"
     for version in VERSIONS:
         ACCOUNTS = load_config["inference"]["versions"][version]["registries"]
         for region in ACCOUNTS.keys():
-            uri = get_huggingface_llm_image_uri("huggingface", region=region, version=version)
+            uri = get_huggingface_llm_image_uri(backend, region=region, version=version)
             expected = expected_uris.huggingface_llm_framework_uri(
                 "huggingface-pytorch-tgi-inference",
                 ACCOUNTS[region],
                 version,
-                HF_VERSIONS_MAPPING[version],
+                HF_VERSIONS_MAPPING[device][version],
                 region=region,
             )
             assert expected == uri